refactor: vo 패키지 위치조정
