+++ src/main/kotlin/domain/Coordinate.kt
@@ -1,6 +1,6 @@
 package domain
 
-import vo.Point
+import domain.vo.Point
 
 data class Coordinate(
     val x: Point,
