+++ src/test/kotlin/domain/vo/HeightTest.kt
@@ -1,20 +1,20 @@
-package vo
+package domain.vo
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FreeSpec
 
-internal class MineCountTest : FreeSpec({
+internal class HeightTest : FreeSpec({
 
     "에러" - {
-        "지뢰개수는 음수가 될 수 없다" {
+        "높이는 음수가 될 수 없다" {
             shouldThrow<IllegalArgumentException> {
-                MineCount(-1)
+                Height(-1)
             }
         }
 
-        "지뢰개수는 0 이 될 수 없다" {
+        "높이는 0 이 될 수 없다" {
             shouldThrow<IllegalArgumentException> {
-                MineCount(-0)
+                Height(-0)
             }
         }
     }
@@ -22,7 +22,7 @@ internal class MineCountTest : FreeSpec({
     "성공" - {
         "1 이상의 숫자를 입력하면 성공한다" {
             val value = (1..100).random()
-            MineCount(value)
+            Height(value)
         }
     }
 })
